+++ include/vkui/Widgets.h
@@ -2,8 +2,10 @@
 
 #pragma once
 
+#include <vkui/widgets/VkComboBox.h>
 #include <vkui/widgets/VkControlSize.h>
 #include <vkui/widgets/controls/VkSegmentedControl.h>
 #include <vkui/widgets/controls/VkSwitch.h>
 #include <vkui/widgets/overlays/VkPopover.h>
 #include <vkui/widgets/style/VkStyle.h>
+#include <vkui/widgets/style/VkStyleSheet.h>